Python script to create a bin file
A BIN file is created that can be used to directly upload to the CSI CAT28C16AP using Xgpro V 10.05 and the XGecu Pro EEPROM programmer. I had to use this because the Ben Eater Arduino EEPROM programmer wasn't working after sometime (initially it was working, but once I tried to program the flag instructions it stopped for some unknown reason).